fix: move CORE_LATEX_TOOLS/IMAGE_TOOLS to shared to break a subsystem cycle
Cursor Bugbot correctly flagged that the previous commit's satisfies check
in src/latex/latexToolchain.ts (a `latex` subsystem file) value-importing
CORE_LATEX_TOOLS from src/tools/setup/toolProbing.ts (a `tools` subsystem
file) introduced a new latex -> tools edge. The `tools -> latex` edge
already exists, so this closed a real cross-subsystem cycle and correctly
failed the LAY-1 subsystem edge ratchet test (config/ratchets/
architecture-edges-baseline.json has no latex -> tools entry in any form).

CORE_LATEX_TOOLS and IMAGE_TOOLS move to src/shared/constants/latex.ts,
which `tools`, `latex`, and `controllers` are all already permitted to
depend on (all three -> shared edges are pre-existing `value` entries in
the baseline) — the correct home for a constant three independent
subsystems need, none of which should depend on each other for it.
toolProbing.ts, ProbeEnvironmentTool.ts, VerifySetupTool.ts,
LatexToolingController.ts, and latexToolchain.ts are updated to import
from the new location; no runtime behavior changes anywhere, and the
LAY-1 ratchet test now passes again.
