+++ src/controllers/settingsView/LatexToolingController.ts
@@ -1,7 +1,9 @@
 // Local imports - shared constants
 import {
+  CORE_LATEX_TOOLS,
   DEPENDENCY_INSTALL_COMMANDS,
   HOMEBREW_INSTALL_COMMAND,
+  IMAGE_TOOLS,
   SCOOP_INSTALL_COMMAND,
   type OSPlatform,
 } from '@shared/constants/latex';
@@ -9,9 +11,8 @@ import {
   DEFAULT_LATEX_SETTINGS_STATUS,
   type LatexSettingsStatus,
 } from '@shared/schemas/settingsViewMessages';
-import { CORE_LATEX_TOOLS, IMAGE_TOOLS } from '@tools/setup/toolProbing';
 
-// `CORE_LATEX_TOOLS`/`IMAGE_TOOLS` (`src/tools/setup/toolProbing.ts`) are the
+// `CORE_LATEX_TOOLS`/`IMAGE_TOOLS` (`@shared/constants/latex`) are the
 // single source of truth for the LaTeX toolchain probe set, shared with
 // `probe_environment`/`verify_setup`. Do not re-list tool names here.
 const LATEX_PROBE_TOOLS = [...CORE_LATEX_TOOLS, ...IMAGE_TOOLS] as const;
